Break-up large `/worlds` command into smaller sub-commands to improve readability and maintenance
