+++ buildsystem-core/src/main/java/com/eintosti/buildsystem/command/subcommand/worlds/BuildersSubCommand.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright (c) 2022, Thomas Meaney
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+package com.eintosti.buildsystem.command.subcommand.worlds;
+
+import com.eintosti.buildsystem.BuildSystem;
+import com.eintosti.buildsystem.command.subcommand.SubCommand;
+import com.eintosti.buildsystem.manager.WorldManager;
+import com.eintosti.buildsystem.object.world.BuildWorld;
+import org.bukkit.entity.Player;
+
+/**
+ * @author einTosti
+ */
+public class BuildersSubCommand implements SubCommand {
+
+    private final BuildSystem plugin;
+    private final String worldName;
+
+    public BuildersSubCommand(BuildSystem plugin, String worldName) {
+        this.plugin = plugin;
+        this.worldName = worldName;
+    }
+
+    @Override
+    public void execute(Player player, String[] args) {
+        WorldManager worldManager = plugin.getWorldManager();
+        if (!worldManager.isPermitted(player, "buildsystem.builders", worldName)) {
+            plugin.sendPermissionMessage(player);
+            return;
+        }
+
+        if (args.length > 2) {
+            player.sendMessage(plugin.getString("worlds_builders_usage"));
+            return;
+        }
+
+        BuildWorld buildWorld = worldManager.getBuildWorld(worldName);
+        if (buildWorld == null) {
+            player.sendMessage(plugin.getString("worlds_builders_unknown_world"));
+            return;
+        }
+
+        plugin.getPlayerManager().getBuildPlayer(player).setCachedWorld(buildWorld);
+        player.openInventory(plugin.getBuilderInventory().getInventory(buildWorld, player));
+    }
+}
